Allow to change settings through the gui
This change allows to configure all parameters (except the log level,
which needs to be fixed at some later point, anyway) through the GUI
under Settings -> PV inverters -> Huawei SUN2000
Note that in order to really uninstall the GUI changes currently a
reboot of the VenusOS is needed. This might be fixed later.